nothing really... ignore this commit

+++ core/src/com/lostmekkasoft/spicewars/PlanetActor.java
@@ -15,7 +15,7 @@ public class PlanetActor extends Actor {
 	TextureRegion textureRegion;
 	float actorX;
 	float actorY;
-	int textureFactor = 2;
+	float textureFactor = 2f;
 
 	public Planet planet;
